+++ test/HttpTaskManagerEpicsTest.java
@@ -1,3 +1,6 @@
+import HttpTaskServer.HttpTaskServer;
+import HttpTaskServer.LocalDateTimeAdapter;
+import HttpTaskServer.DurationAdapter;
 import com.google.gson.*;
 import manager.InMemoryTaskManager;
 import manager.TaskManager;
@@ -21,7 +24,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 class HttpTaskManagerEpicsTest {
     // создаём экземпляр InMemoryTaskManager
     TaskManager manager = new InMemoryTaskManager();
-    // передаём его в качестве аргумента в конструктор HttpTaskServer
+    // передаём его в качестве аргумента в конструктор HttpTaskServer.HttpTaskServer
     HttpTaskServer taskServer = new HttpTaskServer(manager);
     Gson gson = new GsonBuilder().registerTypeAdapter(LocalDateTime.class, new LocalDateTimeAdapter())
             .registerTypeAdapter(Duration.class, new DurationAdapter())
